Refactor to use functions and to use a dedicated accounting db account
