+++ OpenStack-accounting/etc/thecount/thecount.conf.example
@@ -1,0 +1,2 @@
+[database]
+connection =mysql+pymysql://<dbusername>:<dbpassword>@<dbhost>:3306
