+++ OpenStack-accounting/readme.md
@@ -1,0 +1,7 @@
+Prerequisites
+python-ConfigParser
+python3-requests
+python36-sqlalchemy
+python36-PyMySQL
+
+Installation
